+++ projects/ng-virtual-list/src/lib/components/ng-list-item/base/base-virtual-list-item-component.ts
@@ -13,6 +13,8 @@ import {
 } from '../const';
 import { TextDirection, TextDirections } from '../../../enums';
 import { NgVirtualListPublicService } from '../../../ng-virtual-list-public.service';
+import { NgVirtualListService } from '../../../ng-virtual-list.service';
+import { createDisplayId } from '../utils';
 
 /**
  * BaseVirtualListItemComponent
@@ -24,6 +26,8 @@ export class BaseVirtualListItemComponent {
   private _cdr = inject(ChangeDetectorRef);
 
   private _apiService = inject(NgVirtualListPublicService);
+
+  protected readonly _service = inject(NgVirtualListService);
 
   protected _id!: number;
   get id() {
@@ -125,6 +129,10 @@ export class BaseVirtualListItemComponent {
   protected _destroyRef = inject(DestroyRef);
 
   constructor() {
+    this._id = this._service.generateComponentId();
+    this._listId = this._service.id;
+    this._displayId = createDisplayId(this._listId, this._id);
+
     this.classes = computed(() => {
       const data = this.data(), focused = this.focused();
       return {